frontend: Add loading state and animation
This commit only introduces a general loading state that the frontend
application enters until the system information could be read from the
WASM binary. This however does not include the time it takes to load the
actual WASM binary. This change will be introduced in a follow-up
commit.
